refactor: Player and follow components
