+++ frontend/js/src/player/PlayerPage.tsx
