+++ frontend/js/src/user/year-in-music/2023/YearInMusic2023.tsx
@@ -45,11 +45,11 @@ import { getEntityLink } from "../../stats/utils";
 import ImageShareButtons from "./components/ImageShareButtons";
 
 import ListenCard from "../../../listens/ListenCard";
-import UserListModalEntry from "../../../follow/UserListModalEntry";
+import UserListModalEntry from "../../components/follow/UserListModalEntry";
 import { JSPFTrackToListen } from "../../../playlists/utils";
 import CustomChoropleth from "../../stats/components/Choropleth";
 import { ToastMsg } from "../../../notifications/Notifications";
-import FollowButton from "../../../follow/FollowButton";
+import FollowButton from "../../components/follow/FollowButton";
 
 export type YearInMusicProps = {
   user: ListenBrainzUser;
